Initial Commit
Notes:
Code is definitely a bit rough around the edges but gets the job done.

Will have to test speed/optimisation/GPU support when I am on my home PC (this was trained in about 5 mins on a dual core 2Ghz processor for 500 episodes)

Any further improvements will have to be researched.